+++ apps/admin-demo/src/components/pdk/DemoSelectInput.vue
@@ -2,6 +2,7 @@
   <select
     :id="id"
     v-model="model"
+    v-test="[AdminComponent.SelectInput, element]"
     :name="id"
     :disabled="options.length === 1 || element.isDisabled || element.isSuspended"
     class="border flex-grow px-2 py-1 rounded">
@@ -14,12 +15,18 @@
   </select>
 </template>
 
-<script lang="ts" setup>
-import {type ElementInstance, type OptionsProp, useSelectInputContext} from '@myparcel-pdk/admin';
+<script generic="T extends SelectInputModelValue" lang="ts" setup>
+import {
+  AdminComponent,
+  type SelectInputEmits,
+  type SelectInputModelValue,
+  type SelectInputProps,
+  useSelectInputContext,
+} from '@myparcel-pdk/admin';
 
 // eslint-disable-next-line vue/no-unused-properties
-const props = defineProps<{element: ElementInstance<OptionsProp>; modelValue: string | number}>();
-const emit = defineEmits<(e: 'update:modelValue', value: string | number) => void>();
+const props = defineProps<SelectInputProps<T>>();
+const emit = defineEmits<SelectInputEmits<T>>();
 
 const {id, model, options} = useSelectInputContext(props, emit);
 </script>