+++ apps/admin-preset-bootstrap4/src/components/Bootstrap4FormGroup.vue
@@ -1,5 +1,7 @@
 <template>
-  <div class="form-group row">
+  <div
+    v-if="isInteractive"
+    class="form-group row">
     <label
       :class="{
         required: !element.isOptional,
@@ -33,14 +35,31 @@
       </div>
     </div>
   </div>
+
+  <div
+    v-else
+    class="form-group row">
+    <div class="col-12">
+      <slot />
+    </div>
+  </div>
 </template>
 
 <script lang="ts" setup>
-import {type ElementInstance, generateFieldId, useLanguage} from '@myparcel-pdk/admin';
+import {computed, toRefs} from 'vue';
+import {type FormGroupProps, generateFieldId, useLanguage} from '@myparcel-pdk/admin';
+import {type InteractiveElementInstance} from '@myparcel/vue-form-builder';
+import {isOfType} from '@myparcel/ts-utils';
 
-const props = defineProps<{element: ElementInstance<{description: string}>}>();
+const props = defineProps<FormGroupProps>();
+
+const propRefs = toRefs(props);
 
 const {translate} = useLanguage();
 
-const id = generateFieldId(props.element);
+const id = generateFieldId(propRefs.element);
+
+const isInteractive = computed<boolean>(() => {
+  return isOfType<InteractiveElementInstance>(propRefs.element, 'ref');
+});
 </script>