+++ libs/admin/core/src/forms/helpers/triStateFieldIsEnabled.ts
@@ -1,11 +1,10 @@
 import {get} from '@vueuse/core';
 import {type FormInstance} from '@myparcel/vue-form-builder';
-import {type ElementInstance} from '../../types';
+import {type ElementInstance, type TriStateInputProps} from '../../types';
 import {TriState} from '../../data';
-import {type TriStateElementProps} from '../../composables';
 
 export const triStateFieldIsEnabled = (form: FormInstance, fieldName: string): boolean => {
-  const field = form.getField(fieldName) as ElementInstance<TriStateElementProps> | undefined;
+  const field = form.getField(fieldName) as unknown as ElementInstance<TriStateInputProps> | undefined;
 
   if (!field) {
     return false;